refactor: move dashboard components to ui/cards directory and remove unused Progress component
