+++ src/components/modules/dashboard/ui/cards/MetricCard.tsx
@@ -1,5 +1,5 @@
 import { cn } from "@/lib/utils";
-import { Card, CardContent } from "../../ui/card";
+import { Card, CardContent } from "../../../../ui/card";
 
 interface MetricCardProps {
   title: string;
